php7-ext: resolve segfault in mapi_icaltomapi
In PHP7, zend_parse_parameters requires that the "s" format specifier
is matched by a size_t pointer (PHP5: int pointer).

+++ php7-ext/main.cpp
@@ -7753,7 +7753,7 @@ ZEND_FUNCTION(mapi_icaltomapi)
 	zval *resAddrBook;
 	zval *resMessage;
 	zend_bool *noRecipients;
-	ULONG cbString = 0;
+	size_t cbString = 0;
 	char *szString = nullptr;
 	IMAPISession *lpMAPISession = nullptr;
 	IAddrBook *lpAddrBook = nullptr;
